Warnings solving commit
Отработаны варнинги

+++ src/physics/sphere/Sphere2D.java
@@ -27,7 +27,7 @@ public abstract class Sphere2D implements Drawable {
 
     @Override
     public void draw(Graphics g) {
-        int[] coords = new int[]{(int) Math.round(x0 - (double) r), (int) Math.round(y0 - (double) r), (int) Math.round(r * 2)};
+        int[] coords = new int[]{(int) Math.round(x0 - (double) r), (int) Math.round(y0 - (double) r), Math.round(r * 2)};
         g.drawOval(coords[0], coords[1], coords[2], coords[2]);
     }
